rename: 세그먼트 이름 단수화
