+++ src/entities/auth/model/LoginCredential.ts
